Move helpers for marshaller shape into separate class
