+++ src/libraries/System.Runtime.InteropServices/gen/Microsoft.Interop.SourceGeneration/Marshalling/ICustomTypeMarshallingStrategy.cs
@@ -74,7 +74,7 @@ namespace Microsoft.Interop
                 InvocationExpression(
                     MemberAccessExpression(SyntaxKind.SimpleMemberAccessExpression,
                         _marshallerTypeSyntax,
-                        IdentifierName(ShapeMemberNames.Value.Stateless.Free)),
+                        IdentifierName(ShapeMemberNames.Free)),
                     ArgumentList(SingletonSeparatedList(
                         Argument(IdentifierName(context.GetIdentifiers(info).native))))));
         }
